IGNITE-28819 Move GridCompoundFuture and NodeStoppingException to ignite-commons
Introduce ExpectedFailure marker in ignite-commons so GridCompoundFuture
can classify benign failures without referencing core exception types.
ClusterTopologyCheckedException, IgniteConsistencyViolationException,
IgniteTxOptimisticCheckedException and IgniteFutureCancelledCheckedException
implement it. Failure-handling behavior is preserved.
