+++ modules/commons/src/main/java/org/apache/ignite/internal/NodeStoppingException.java
