+++ modules/core/src/main/java/org/apache/ignite/internal/cluster/ClusterTopologyCheckedException.java
@@ -18,13 +18,14 @@
 package org.apache.ignite.internal.cluster;
 
 import org.apache.ignite.IgniteCheckedException;
+import org.apache.ignite.internal.ExpectedFailure;
 import org.apache.ignite.internal.IgniteInternalFuture;
 import org.jetbrains.annotations.Nullable;
 
 /**
  * This exception is used to indicate error with grid topology (e.g., crashed node, etc.).
  */
-public class ClusterTopologyCheckedException extends IgniteCheckedException {
+public class ClusterTopologyCheckedException extends IgniteCheckedException implements ExpectedFailure {
     /** */
     private static final long serialVersionUID = 0L;
